info/whisper_py/requirements.txt
torch>=2.1.0
openai-whisper
langchain-core